+++ CHANGELOG.md
@@ -17,6 +17,7 @@
 ### 💡 Enhancements 💡
 - `pkg/translator/prometheusremotewrite`: Allow to disable sanitize metric labels (#8270)
 - `basicauthextension`: Implement `configauth.ClientAuthenticator` so that the extension can also be used as HTTP client basic authenticator.(#8847)
+- `azuremonitorexporter`, `lokiexporter`, `observiqexporter`: Update timestamp processing logic (#9130)
 
 ### 🧰 Bug fixes 🧰